db/db_requirements.txt
dbt-core
dbt-postgres
psycopg2-binary
python-dotenv
xarray
sqlalchemy
fastkml
lxml
pykml
bs4